requirements.txt
asgiref==3.8.1
Django==3.0
django-nested-admin==3.4.0
pillow==10.3.0
python-decouple==3.8
python-monkey-business==1.0.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
